requirements.txt
blinker==1.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.45.1
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.5
MarkupSafe==3.0.2
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.10
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
secure-smtplib==0.1.1
six==1.16.0
urllib3==2.2.0
Werkzeug==3.1.3
python-binance
cryptography
ccxt
aiohttp
